Update the base header file for the verbs API
ib_verbs.h defines the basic types and functions for verbs API.
But for RDMA para-virtualization, most of the verbs calls are forwarded
to the host backend, so we don't need most of the verbs implementations
and declarations. Here we only keep the minimum types definitions and
functions declarations, and the rests are commented out, which should
be cleaned up later.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>