Update the verbs implementation file
verbs.cc contains the implementatoins of the verbs API. All the functions
have been declared in ib_verbs.h header file. Same as the last commit,
we don't need all the verbs implementation, but only the necessary ones
and the basic internal support. All the rests are commented out for later
clean up.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>