Update the initialization functions for the virtual driver
Due to the nature of OSv, we don't need to distinguish the
user and kernel space. And as it is a virtual driver, it is
 not necessary to have exact the same structure as the real one.

The basic idea here it to put all device information in one
place, i.e. vrdma.config, which contains both kernel and user
drivers info/properties. The advatanges of doing this are:
we don't need sysfs anymore; it will be much easier to configure
the properties of the virtual driver; no single pieces of files
flying around in the system, e.g. in /sys/class/infiniband_verbs.

Rewrite the ibv_read_sysfs_file function, and read only on the
vrdma.config file.

Clean up the unused stuffs.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>