Update Makefiles to compile both C and C++ source files
As OSv is C and C++ mixed, when we need to compile a C library
with OSv's C++ files, we have to compile this file also with C++
compiler to avoid a lot of conflicts.

This update will add gnu++11 for C++ source files and compile
them as objects first, and then add gnu99 for C source files
and link the C++ objects to generate the output shared library.

Move the mlx4 library to the directory where the
libibverbs.so is, so we don't have dynamic link problems.

Add libibverbs include dir as the include path for the other
two libraries.

Use BEGIN/END_C_DECL to protect the files that will be used
as both C and C++.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>