The virtio-rdma frontend driver
This is the main virtio-rdma implementation, which implements hypercall
using OSv internal support. This implementation is compatible with HyV
backend driver. Handlers and basic data types/structures have been somehow rewritten
according to the OSv implementation.

Hypercall event sequence is different to the original Hyv, as we are not
able to initialize the vring in pci probe phase in OSv, but doing this in
the pci device initialization phase should also work.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>