Add support for initialize the memory region
Update the vrdma_alloc_ucontext with virtual context support.

Add vrdma_reg_mr for initializing the memory regions in the guest.
The allocated memory region is contigious to the guest application,
but it is allocated in single page chunks at moment. These chunks
will be mapped and handled on the host side accordingly for the
further operation.

The deallocation on failure will be added later.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>