Fix the problem of initializing udata
For the original OFED, verb commands from user space
have to be copied over to the kernel space and restructured.
The input buffer for the kernel verb call is at the end of
the command buffer, and the output buffer is in the middle
of the response member of the command buffer.

This commit changes the way of initializing the udata object,
which takes the command parameter to the device. As we are using
OSv, there is no need to copy the command over and over, and no
need to make many new buffers, just use the one that passed from
the user verbs API. This will apply also for the further verbs
call support.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>