Fix the buffer mapping problem between HyV and IB
The ibudata structure acturally hides a tail buffer that is used
by the mlx4 provider, and it is not exposed to the kernel driver.
We wanted to avoid several memcpy calls by redirect the in/out
buffer of the ibudata to the HyV defined communication buffer,
but at moment it's not going to work, as ibudata is allocated
contiguously, and it has a different structrue to HyV buffer.

At moment, the easy solution is just keeping the memcpy as is,
and will find a better way later on.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>