Fix the memory mapping problem
This problem was because of the replacement of get_user_pages
function. As the memory that malloc'ed by OSv is already pinned,
we don't need to do it again. And malloc will try to allocate
contiguous memory by default. This makes things much easier.
We just make one chunk of the translation, add pass the guest
physical address to host. On the host side, the backend driver
will remap everything into pages and map them with the RDMA
network card.

In the case of large memory request, if OSv is not able to handle
such large and contiguous memory region, this fix will not work.
We can still live with this fix for a while, until this very special
case happens.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>