Fix the doorbell problem for vRDMA on OSv
In the original HyV implementation, libmlx4 creates the doorbell
with mmap call, and the HyV driver allocates a new region of memory
and remaps it to the virtual address that libmlx4 generates. This
new doorbell memory address from HyV, not the libmlx4 one, will be
translated and passed to the host.

Fot OSv, it's a bit different, as mmap'ed memory address is in a
special space, and thus cannot be mapped to physical address space.
The original logic will not work unless we replace the mmap and do
the same remap.  As we have the full control of libmlx4, we don't
really need to mmap the doorbell in libmlx4, but rather in the
virtio-rdma driver, and the virtio-rdma driver will just malloc
the doorbell and then translate it to the host. The good things
are: 1. we don't need to allocate the same memory twice; 2. we
can avoid using mmap; 3. no remap is necessary.

A few function calls have been updated in order to pass the
doorbeel address between libmlx4 and virtio-rdma.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>