Fix the timing issue with mutex locks
We have to make sure that after kicking the virtqueue, the main
thread waits until the element is processed by the backend driver,
and the virtqueue element is actually used.

The original driver implemented the lock with spin_lock_irqrestore
and spin_unlock_irqrestore, which are not available on OSv. Here
we re-implement the same lock mechanism using mutex and conditional
variables.

Signed-off-by: Shiqing Fan <shiqing.fan@huawei.com>